requirements.txt
asgiref==3.5.2
Django==4.1.2
markdown2==2.4.6
sqlparse==0.4.3
tzdata==2022.5
